NPZ.jl

NPZ.npzread — Method

npzread(filename::AbstractString, [vars])

Read a variable or a collection of variables from filename. The input needs to be either an npy or an npz file. The optional argument vars is used only for npz files. If it is specified, only the matching variables are read in from the file.

Example

julia> npzwrite("temp.npz", Dict("x" => ones(3), "y" => 3))

julia> npzread("temp.npz") # Reads all variables
Dict{String,Any} with 2 entries:
  "x" => [1.0, 1.0, 1.0]
  "y" => 3

julia> npzread("temp.npz", ["x"]) # Reads only "x"
Dict{String,Array{Float64,1}} with 1 entry:
  "x" => [1.0, 1.0, 1.0]

NPZ.npzwrite — Method

npzwrite(filename::AbstractString, x)

Write the variable x to the npy file filename. Unlike numpy, the extension .npy is not appened to filename.

Warning

Any existing file with the same name will be overwritten.

Examples

julia> npzwrite("abc.npy", zeros(3))

julia> npzread("abc.npy")
3-element Array{Float64,1}:
 0.0
 0.0
 0.0

NPZ.npzwrite — Method

npzwrite(filename::AbstractString, vars::Dict{<:AbstractString})

Write the variables in vars to an npz file named filename. Unlike numpy, the extension .npz is not appened to filename.

Warning

Any existing file with the same name will be overwritten.

julia> npzwrite("temp.npz", Dict("x" => ones(3), "y" => 3))

julia> npzread("temp.npz")
Dict{String,Any} with 2 entries:
  "x" => [1.0, 1.0, 1.0]
  "y" => 3